rpi4: Rename files and modules to be less generic
The rpi4 code uses very generic names (genet, as an example) for
headers and modules. Make these less generic to avoid accidental
clashes in the future. The new module names use underscores (most
relevant for network drivers where hyphens are not allowed) and
spell out trailing numbers.

While here, clean up `sys/gpio.h` and remove it from `virt`, where
it is unused.
