+++ usr/src/psm/stand/boot/aarch64/rpi4/rpi4.c
@@ -39,8 +39,8 @@
 #include <sys/gpio.h>
 #include <sys/bcm2835_mbox.h>
 #include <sys/bcm2835_mboxreg.h>
-#include <sys/vcprop.h>
-#include <sys/vcio.h>
+#include <sys/bcm2835_vcprop.h>
+#include <sys/bcm2835_vcio.h>
 
 static void
 cache_flush(void *addr, size_t len)
