+++ usr/src/uts/armv8/rpi4/Makefile.rpi4
@@ -40,8 +40,8 @@ ROOT_PLAT_DRV_DIRS	= $(ROOT_RPI4_DRV_DIR)
 # Define modules
 #
 RPI4_KMODS	=	\
-	genet		\
-	bcm2711-emmc2
+	bcm2711_emmctwo	\
+	bcm2711_genet
 
 include $(UTSBASE)/armv8/rpi4/Makefile.files
