+++ usr/src/uts/armv8/rpi4/bcm2711_genet/Makefile
@@ -31,8 +31,8 @@ UTSBASE	= ../../..
 #
 #	Define the module and object file sets.
 #
-MODULE		= genet
-OBJECTS		= $(GENET_OBJS:%=$(OBJS_DIR)/%)
+MODULE		= bcm2711_genet
+OBJECTS		= $(BCM2711_GENET_OBJS:%=$(OBJS_DIR)/%)
 ROOTMODULE	= $(ROOT_RPI4_DRV_DIR)/$(MODULE)
 CONF_SRCDIR	= $(UTSBASE)/armv8/rpi4/io/$(MODULE)/
