+++ usr/src/uts/armv8/rpi4/io/bcm2711_genet/bcm2711_genet.conf
