+++ usr/src/uts/armv8/rpi4/io/bcm2711_genet/bcm2711_genet.h
@@ -36,7 +36,7 @@ extern "C" {
 #include <sys/ethernet.h>
 #include <sys/sunddi.h>
 #include <sys/ddi_impldefs.h>
-#include <sys/genetreg.h>
+#include <sys/bcm2711_genetreg.h>
 
 struct genet_reg {
 	ddi_acc_handle_t handle;
