+++ usr/src/uts/armv8/rpi4/io/rpi4.c
@@ -42,8 +42,8 @@
 #include <vm/hat.h>
 #include <sys/bcm2835_mbox.h>
 #include <sys/bcm2835_mboxreg.h>
-#include <sys/vcprop.h>
-#include <sys/vcio.h>
+#include <sys/bcm2835_vcprop.h>
+#include <sys/bcm2835_vcio.h>
 #include <sys/gpio.h>
 
 /*
