--- usr/src/uts/armv8/virt/sys/gpio.h
@@ -1,45 +1,0 @@
-/*
- * CDDL HEADER START
- *
- * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
- *
- * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
- * or http://www.opensolaris.org/os/licensing.
- * See the License for the specific language governing permissions
- * and limitations under the License.
- *
- * When distributing Covered Code, include this CDDL HEADER in each
- * file and include the License file at usr/src/OPENSOLARIS.LICENSE.
- * If applicable, add the following below this CDDL HEADER, with the
- * fields enclosed by brackets "[]" replaced with your own identifying
- * information: Portions Copyright [yyyy] [name of copyright owner]
- *
- * CDDL HEADER END
- */
-/*
- * Copyright 2017 Hayashi Naoyuki
- */
-
-#ifndef _SYS_GPIO_H
-#define _SYS_GPIO_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#include <sys/promif.h>
-#include <sys/types.h>
-
-struct gpio_ctrl {
-	pnode_t node;
-	int pin;
-};
-
-#ifdef	__cplusplus
-}
-#endif
-
-#endif	/* _SYS_GPIO_H */
